Extend juicefs encript options support, To #51793350
